refactor(user-service): extract SessionService from AuthService + add .dockerignore for all services
